requirements.txt
# 🚀 PS-05 Document Understanding System - Core Dependencies

# Core ML Framework
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Computer Vision
opencv-python>=4.8.0
Pillow>=9.0.0
albumentations>=1.3.0

# Document Processing
PyMuPDF>=1.23.0
python-docx>=0.8.11
python-pptx>=0.6.21

# OCR & Text Processing
easyocr>=1.7.0
pytesseract>=0.3.10
langdetect>=1.0.9

# Natural Language Processing
transformers>=4.30.0
tokenizers>=0.13.0
sentencepiece>=0.1.99

# YOLO & Object Detection
ultralytics>=8.0.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Utilities
tqdm>=4.65.0
PyYAML>=6.0
click>=8.1.0

# Evaluation & Metrics
pycocotools>=2.0.6
jiwer>=3.0.0
sacrebleu>=2.3.0
bert-score>=0.3.11

# Web Framework (Backend)
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6

# Development & Testing
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0

# Optional: GPU Optimization
# nvidia-ml-py>=11.0.0  # Uncomment for GPU monitoring